refactoring attacker and indexing
